backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.0
Flask-Session==0.8.0
mysql-connector-python==9.2.0
PyJWT==2.7.0
cryptography==44.0.2
bcrypt==4.0.1
Werkzeug==2.3.7
Flask-DebugToolbar==0.14.0
pytest==7.4.2
flake8==6.1.0
pyoqs==0.8.0
oqs==0.10.2
python-dotenv==1.0.1